+++ CHANGELOG.md
@@ -9,6 +9,7 @@ All notable changes to this project will be documented in this file.
 
 - OAuth 2.0 using the client credentials token flow is now supported by the `http_client` input and output, and the `http` processor.
 - Method `format_timestamp` added to Bloblang.
+- Bloblang now supports trailing commas in object and array literals and function and method parameters.
 
 ### Fixed